Refactor: 프로젝트 디렉토리 정리
- RedisConfig: RedisTemplate Bean 메서드 오타 수정
- RefreshToken: global.jwt -> global.type 디렉토리 이동
- RandomNicknameGenerator -> RandomCodeGenerator 클래스명 수정
