+++ server/src/main/java/site/bannabe/server/global/jwt/GenerateToken.java
@@ -1,5 +1,7 @@
 package site.bannabe.server.global.jwt;
 
+import site.bannabe.server.global.type.RefreshToken;
+
 public record GenerateToken(
     String accessToken,
     RefreshToken refreshToken
