+++ server/src/main/java/site/bannabe/server/global/type/RefreshToken.java
@@ -1,4 +1,4 @@
-package site.bannabe.server.global.jwt;
+package site.bannabe.server.global.type;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
